pyproject.toml
[project]
name = "coach-conejito"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "altair<6",
    "fitparse>=1.2.0",
    "garminconnect>=0.2.8",
    "google-generativeai>=0.8.6",
    "mlx-lm>=0.29.1",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "watchdog>=6.0.0",
]
